Split tests by language
